+++ baseapp/abci.go
@@ -230,8 +230,7 @@ func (app *BaseApp) BeginBlock(req abci.RequestBeginBlock) (res abci.ResponseBeg
 
 	// call the hooks with the BeginBlock messages
 	for _, streamingListener := range app.abciListeners {
-		goCtx := sdk.WrapSDKContext(app.deliverState.ctx)
-		if err := streamingListener.ListenBeginBlock(goCtx, req, res); err != nil {
+		if err := streamingListener.ListenBeginBlock(app.deliverState.ctx, req, res); err != nil {
 			panic(fmt.Errorf("BeginBlock listening hook failed, height: %d, err: %w", req.Header.Height, err))
 		}
 	}
@@ -260,8 +259,7 @@ func (app *BaseApp) EndBlock(req abci.RequestEndBlock) (res abci.ResponseEndBloc
 
 	// call the streaming service hooks with the EndBlock messages
 	for _, streamingListener := range app.abciListeners {
-		goCtx := sdk.WrapSDKContext(app.deliverState.ctx)
-		if err := streamingListener.ListenEndBlock(goCtx, req, res); err != nil {
+		if err := streamingListener.ListenEndBlock(app.deliverState.ctx, req, res); err != nil {
 			panic(fmt.Errorf("EndBlock listening hook failed, height: %d, err: %w", req.Height, err))
 		}
 	}
@@ -317,7 +315,7 @@ func (app *BaseApp) DeliverTx(req abci.RequestDeliverTx) (res abci.ResponseDeliv
 	defer func() {
 		for _, streamingListener := range app.abciListeners {
 			if err := streamingListener.ListenDeliverTx(app.deliverState.ctx, req, res); err != nil {
-				app.logger.Error("DeliverTx listening hook failed", "err", err)
+				panic(fmt.Errorf("DeliverTx listening hook failed: %w", err))
 			}
 		}
 	}()
@@ -354,8 +352,6 @@ func (app *BaseApp) DeliverTx(req abci.RequestDeliverTx) (res abci.ResponseDeliv
 // against that height and gracefully halt if it matches the latest committed
 // height.
 func (app *BaseApp) Commit() abci.ResponseCommit {
-	defer telemetry.MeasureSince(time.Now(), "abci", "commit")
-
 	header := app.deliverState.ctx.BlockHeader()
 	retainHeight := app.GetBlockRetentionHeight(header.Height)
 
@@ -372,8 +368,7 @@ func (app *BaseApp) Commit() abci.ResponseCommit {
 
 	// call the hooks with the Commit message
 	for _, streamingListener := range app.abciListeners {
-		goCtx := sdk.WrapSDKContext(app.deliverState.ctx)
-		if err := streamingListener.ListenCommit(goCtx, res); err != nil {
+		if err := streamingListener.ListenCommit(app.deliverState.ctx, res); err != nil {
 			panic(fmt.Errorf("Commit listening hook failed, height: %d, err: %w", header.Height, err))
 		}
 	}
@@ -407,10 +402,7 @@ func (app *BaseApp) Commit() abci.ResponseCommit {
 		app.halt()
 	}
 
-	var snapshotHeight int64
-	if app.snapshotInterval > 0 && uint64(header.Height)%app.snapshotInterval == 0 {
-		snapshotHeight = header.Height
-	}
+	go app.snapshotManager.SnapshotIfApplicable(header.Height)
 
 	return res
 }