+++ store/streaming/file/service.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"io"
 	"os"
 	"path"
 	"sort"
@@ -30,40 +31,29 @@ type StreamingService struct {
 
 	currentBlockNumber int64
 	blockMetadata      types.BlockMetadata
-
-	// outputMetadata, if true, writes additional metadata to file per block
+	// if write the metadata file, otherwise only data file is outputted.
 	outputMetadata bool
-
-	// stopNodeOnErr, if true, will panic and stop the node during ABCI Commit
-	// to ensure eventual consistency of the output, otherwise, any errors are
-	// logged and ignored which could yield data loss in streamed output.
+	// if true, when commit failed it will panic and stop the consensus state machine to ensure the
+	// eventual consistency of the output, otherwise the error is ignored and have the risk of lossing data.
 	stopNodeOnErr bool
-
-	// fsync, if true, will execute file Sync to make sure the data is persisted
-	// onto disk, otherwise there is a risk of data loss during any crash.
+	// if true, the file.Sync() is called to make sure the data is persisted onto disk, otherwise it risks lossing data when system crash.
 	fsync bool
 }
 
-func NewStreamingService(
-	writeDir, filePrefix string,
-	storeKeys []types.StoreKey,
-	cdc codec.BinaryCodec,
-	outputMetadata, stopNodeOnErr, fsync bool,
-) (*StreamingService, error) {
+// NewStreamingService creates a new StreamingService for the provided writeDir, (optional) filePrefix, and storeKeys
+func NewStreamingService(writeDir, filePrefix string, storeKeys []types.StoreKey, c codec.BinaryCodec, outputMetadata bool, stopNodeOnErr bool, fsync bool) (*StreamingService, error) {
 	// sort storeKeys for deterministic output
 	sort.SliceStable(storeKeys, func(i, j int) bool {
 		return storeKeys[i].Name() < storeKeys[j].Name()
 	})
 
-	// NOTE: We use the same listener for each store.
 	listeners := make([]*types.MemoryListener, len(storeKeys))
+	// in this case, we are using the same listener for each Store
 	for i, key := range storeKeys {
 		listeners[i] = types.NewMemoryListener(key)
 	}
-
-	// Check that the writeDir exists and is writable so that we can catch the
-	// error here at initialization. If it is not we don't open a dstFile until we
-	// receive our first ABCI message.
+	// check that the writeDir exists and is writable so that we can catch the error here at initialization if it is not
+	// we don't open a dstFile until we receive our first ABCI message
 	if err := isDirWriteable(writeDir); err != nil {
 		return nil, err
 	}
@@ -72,7 +62,7 @@ func NewStreamingService(
 		storeListeners: listeners,
 		filePrefix:     filePrefix,
 		writeDir:       writeDir,
-		codec:          cdc,
+		codec:          c,
 		outputMetadata: outputMetadata,
 		stopNodeOnErr:  stopNodeOnErr,
 		fsync:          fsync,
@@ -87,65 +77,56 @@ func (fss *StreamingService) Listeners() map[types.StoreKey][]types.WriteListene
 	for _, listener := range fss.storeListeners {
 		listeners[listener.StoreKey()] = []types.WriteListener{listener}
 	}
-
 	return listeners
 }
 
-// ListenBeginBlock satisfies the ABCIListener interface. It sets the received
-// BeginBlock request, response and the current block number. Note, these are
-// not written to file until ListenCommit is executed and outputMetadata is set,
-// after which it will be reset again on the next block.
-func (fss *StreamingService) ListenBeginBlock(ctx context.Context, req abci.RequestBeginBlock, res abci.ResponseBeginBlock) error {
+// ListenBeginBlock satisfies the baseapp.ABCIListener interface
+// It writes the received BeginBlock request and response and the resulting state changes
+// out to a file as described in the above the naming schema
+func (fss *StreamingService) ListenBeginBlock(ctx context.Context, req abci.RequestBeginBlock, res abci.ResponseBeginBlock) (rerr error) {
 	fss.blockMetadata.RequestBeginBlock = &req
 	fss.blockMetadata.ResponseBeginBlock = &res
 	fss.currentBlockNumber = req.Header.Height
 	return nil
 }
 
-// ListenDeliverTx satisfies the ABCIListener interface. It appends the received
-// DeliverTx request and response to a list of DeliverTxs objects. Note, these
-// are not written to file until ListenCommit is executed and outputMetadata is
-// set, after which it will be reset again on the next block.
-func (fss *StreamingService) ListenDeliverTx(ctx context.Context, req abci.RequestDeliverTx, res abci.ResponseDeliverTx) error {
+// ListenDeliverTx satisfies the baseapp.ABCIListener interface
+// It writes the received DeliverTx request and response and the resulting state changes
+// out to a file as described in the above the naming schema
+func (fss *StreamingService) ListenDeliverTx(ctx context.Context, req abci.RequestDeliverTx, res abci.ResponseDeliverTx) (rerr error) {
 	fss.blockMetadata.DeliverTxs = append(fss.blockMetadata.DeliverTxs, &types.BlockMetadata_DeliverTx{
 		Request:  &req,
 		Response: &res,
 	})
-
 	return nil
 }
 
-// ListenEndBlock satisfies the ABCIListener interface. It sets the received
-// EndBlock request, response and the current block number. Note, these are
-// not written to file until ListenCommit is executed and outputMetadata is set,
-// after which it will be reset again on the next block.
-func (fss *StreamingService) ListenEndBlock(ctx context.Context, req abci.RequestEndBlock, res abci.ResponseEndBlock) error {
+// ListenEndBlock satisfies the baseapp.ABCIListener interface
+// It writes the received EndBlock request and response and the resulting state changes
+// out to a file as described in the above the naming schema
+func (fss *StreamingService) ListenEndBlock(ctx context.Context, req abci.RequestEndBlock, res abci.ResponseEndBlock) (rerr error) {
 	fss.blockMetadata.RequestEndBlock = &req
 	fss.blockMetadata.ResponseEndBlock = &res
 	return nil
 }
 
-// ListenCommit satisfies the ABCIListener interface. It is executed during the
-// ABCI Commit request and is responsible for writing all staged data to files.
-// It will only return a non-nil error when stopNodeOnErr is set.
+// ListenEndBlock satisfies the baseapp.ABCIListener interface
 func (fss *StreamingService) ListenCommit(ctx context.Context, res abci.ResponseCommit) error {
-	if err := fss.doListenCommit(ctx, res); err != nil {
+	err := fss.doListenCommit(ctx, res)
+	if err != nil {
 		if fss.stopNodeOnErr {
 			return err
 		}
 	}
-
 	return nil
 }
 
 func (fss *StreamingService) doListenCommit(ctx context.Context, res abci.ResponseCommit) (err error) {
 	fss.blockMetadata.ResponseCommit = &res
 
-	// Write to target files, the file size is written at the beginning, which can
-	// be used to detect completeness.
+	// write to target files, the file size is written at the beginning, which can be used to detect completeness.
 	metaFileName := fmt.Sprintf("block-%d-meta", fss.currentBlockNumber)
 	dataFileName := fmt.Sprintf("block-%d-data", fss.currentBlockNumber)
-
 	if fss.filePrefix != "" {
 		metaFileName = fmt.Sprintf("%s-%s", fss.filePrefix, metaFileName)
 		dataFileName = fmt.Sprintf("%s-%s", fss.filePrefix, dataFileName)
@@ -156,7 +137,6 @@ func (fss *StreamingService) doListenCommit(ctx context.Context, res abci.Respon
 		if err != nil {
 			return err
 		}
-
 		if err := writeLengthPrefixedFile(path.Join(fss.writeDir, metaFileName), bz, fss.fsync); err != nil {
 			return err
 		}
@@ -166,26 +146,32 @@ func (fss *StreamingService) doListenCommit(ctx context.Context, res abci.Respon
 	if err := fss.writeBlockData(&buf); err != nil {
 		return err
 	}
-
 	return writeLengthPrefixedFile(path.Join(fss.writeDir, dataFileName), buf.Bytes(), fss.fsync)
 }
 
 func (fss *StreamingService) writeBlockData(writer io.Writer) error {
 	for _, listener := range fss.storeListeners {
 		cache := listener.PopStateCache()
-
 		for i := range cache {
 			bz, err := fss.codec.MarshalLengthPrefixed(&cache[i])
 			if err != nil {
 				return err
 			}
-
-			if _, err := writer.Write(bz); err != nil {
+			if _, err = writer.Write(bz); err != nil {
 				return err
 			}
 		}
 	}
+	return nil
+}
 
+// Stream satisfies the baseapp.StreamingService interface
+func (fss *StreamingService) Stream(wg *sync.WaitGroup) error {
+	return nil
+}
+
+// Close satisfies the io.Closer interface, which satisfies the baseapp.StreamingService interface
+func (fss *StreamingService) Close() error {
 	return nil
 }
 
@@ -213,30 +199,25 @@ func writeLengthPrefixedFile(path string, data []byte, fsync bool) (err error) {
 	if err != nil {
 		return sdkerrors.Wrapf(err, "open file failed: %s", path)
 	}
-
 	defer func() {
 		// avoid overriding the real error with file close error
 		if err1 := f.Close(); err1 != nil && err == nil {
 			err = sdkerrors.Wrapf(err, "close file failed: %s", path)
 		}
 	}()
-
 	_, err = f.Write(sdk.Uint64ToBigEndian(uint64(len(data))))
 	if err != nil {
 		return sdkerrors.Wrapf(err, "write length prefix failed: %s", path)
 	}
-
 	_, err = f.Write(data)
 	if err != nil {
 		return sdkerrors.Wrapf(err, "write block data failed: %s", path)
 	}
-
 	if fsync {
 		err = f.Sync()
 		if err != nil {
 			return sdkerrors.Wrapf(err, "fsync failed: %s", path)
 		}
 	}
-
-	return err
+	return
 }